flowedit: Extract utils.rs, node_layout.rs, move FlowCanvasState (#2611)
Create utils.rs with pure utility functions: derive_short_name,
split_route, connection_references_node, format_value,
truncate_source, base_port_name, rounded_rect,
check_port_type_compatibility.

Create node_layout.rs with NodeLayout struct, its methods, and
build_from_flow (formerly build_render_nodes) as an associated
function.

Move FlowCanvasState (struct, Default, methods including auto_fit)
from flow_canvas.rs to window_state.rs.

Move tests to their corresponding modules: 25 to utils.rs, 8 to
node_layout.rs, 7 to window_state.rs, keeping 16 in flow_canvas.rs.
